requirements.txt
PyYAML
duckdb
pandas
SpiffWorkflow
shiny
htmltools
# Legacy Streamlit dependencies (keeping for transition period)
streamlit
streamlit_agraph